requirements.txt
crewai
crewai-tools
gradio
openai>=1.0.0
PyPDF2
python-dotenv

faiss-cpu
numpy

tiktoken

sentence-transformers
python-docx
fpdf2
